Move stages under promtail and move / cleanup windows specific target code
