+++ internal/component/loki/source/windowsevent/target.go
@@ -19,10 +19,10 @@ import (
 
 	"github.com/prometheus/prometheus/model/labels"
 
+	"github.com/grafana/alloy/internal/component/loki/source/windowsevent/win_eventlog"
 	"github.com/grafana/alloy/internal/loki/promtail/api"
 	"github.com/grafana/alloy/internal/loki/promtail/scrapeconfig"
 	"github.com/grafana/alloy/internal/loki/promtail/targets/target"
-	"github.com/grafana/alloy/internal/loki/promtail/targets/windows/win_eventlog"
 
 	util_log "github.com/grafana/loki/v3/pkg/util/log"
 )
