+++ internal/component/loki/source/windowsevent/win_eventlog/event.go
