+++ internal/component/loki/source/windowsevent/win_eventlog/syscall_windows.go
