+++ internal/component/loki/source/windowsevent/win_eventlog/util.go
