+++ internal/component/loki/source/windowsevent/win_eventlog/util_test.go
