+++ internal/component/loki/source/windowsevent/win_eventlog/win_eventlog_test.go
