+++ internal/component/loki/source/windowsevent/win_eventlog/zsyscall_windows.go
