+++ internal/loki/promtail/stages/drop.go
