+++ internal/loki/promtail/stages/eventlogmessage.go
