+++ internal/loki/promtail/stages/extensions.go
