+++ internal/loki/promtail/stages/labeldrop.go
