+++ internal/loki/promtail/stages/metrics.go
