+++ internal/loki/promtail/stages/multiline.go
