+++ internal/loki/promtail/stages/output.go
