+++ internal/loki/promtail/stages/regex.go
