+++ internal/loki/promtail/stages/replace.go
