+++ internal/loki/promtail/stages/sampling.go
