+++ internal/loki/promtail/stages/stage.go
