+++ internal/loki/promtail/stages/static_labels.go
