+++ internal/loki/promtail/stages/template.go
