+++ internal/loki/promtail/stages/tenant.go
