+++ internal/loki/promtail/stages/timestamp.go
